Fix imports after moving Space class
