+++ src/main/control/Human.java
@@ -1,7 +1,7 @@
 package main.control;
 
-import main.Space;
 import main.boards.Board;
+import main.boards.Space;
 import main.moves.Move;
 import main.moves.PromotePawnMove;
 import main.pieces.Piece;
