+++ src/main/control/Player.java
@@ -3,8 +3,8 @@ package main.control;
 import main.moves.CastlingMove;
 import main.moves.Move;
 import main.moves.StandardMove;
-import main.Space;
 import main.boards.Board;
+import main.boards.Space;
 import main.pieces.King;
 
 import java.util.ArrayList;
