+++ src/main/pieces/King.java
@@ -3,8 +3,8 @@ package main.pieces;
 import main.moves.CastlingMove;
 import main.moves.Move;
 import main.moves.StandardMove;
-import main.Space;
 import main.boards.Board;
+import main.boards.Space;
 
 import java.util.ArrayList;
 import java.util.List;
