+++ src/main/pieces/Pawn.java
@@ -5,8 +5,8 @@ import main.moves.EnPassantMove;
 import main.moves.Move;
 import main.moves.PromotePawnMove;
 import main.moves.StandardMove;
-import main.Space;
 import main.boards.Board;
+import main.boards.Space;
 
 import java.util.ArrayList;
 import java.util.List;
