+++ src/main/pieces/Piece.java
@@ -1,8 +1,8 @@
 package main.pieces;
 
 import main.moves.Move;
-import main.Space;
 import main.boards.Board;
+import main.boards.Space;
 
 import java.util.ArrayList;
 import java.util.List;
